% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{check_constraints}
\alias{check_constraints}
\title{Check constraint violation in ODE}
\usage{
check_constraints(constraints, envir, current_time)
}
\arguments{
\item{constraints}{Vector of strings, specifying the statements to evaluate, as returned by get_logical_constraints()}

\item{envir}{Environment}

\item{current_time}{Current time in ODE}
}
\value{
Nothing if constraints are not violated; otherwise a message will be printed with the constraints that were violated, and the simulation will terminate
}
\description{
Check constraint violation in ODE
}
