% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_name}
\alias{clean_name}
\title{Clean variable name(s)}
\usage{
clean_name(new, protected = NULL)
}
\arguments{
\item{new}{Vector of names to transform to valid names}

\item{protected}{Optional vector of protected names, e.g., existing names in model}
}
\value{
Vector of cleaned names
}
\description{
Clean variable name(s) to create syntactically valid, unique names for use in R and Julia.
}
\examples{
sfm <- sdbuildR("predator_prey")
# As the variable name "predator" is already taken, clean_name() will create
# a unique name
clean_name("predator", as.data.frame(sfm)[["name"]])

}
\concept{internal}
