% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compare_sim}
\alias{compare_sim}
\title{Compare two simulations}
\usage{
compare_sim(sim1, sim2, tolerance = 1e-05)
}
\arguments{
\item{sim1}{Simulation 1}

\item{sim2}{Simulation 2}

\item{tolerance}{Numeric; tolerance for comparing values. Defaults to 0.00001.}
}
\value{
List with comparison results
}
\description{
Compare two simulations
}
\seealso{
Other internal: 
\code{\link{has_internet}()},
\code{\link{nonnegative}()},
\code{\link{not_on_cran}()},
\code{\link{saveat_func}()}
}
\concept{internal}
