% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdbuildR_custom_func.R
\name{contains_IM}
\alias{contains_IM}
\title{Check if needle is in haystack}
\usage{
contains_IM(haystack, needle)
}
\arguments{
\item{haystack}{Vector or string to search through}

\item{needle}{Value to search for}
}
\value{
Logical value
}
\description{
Check whether value is in vector or string. Equivalent of .Contains() in Insight Maker.
}
\examples{
contains_IM(c("a", "b", "c"), "d") # FALSE
contains_IM(c("abcdef"), "bc") # TRUE
}
\concept{custom}
