% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_unit.R
\name{custom_unit}
\alias{custom_unit}
\title{Create or modify custom units}
\usage{
custom_unit(object, name, eqn = 1, doc = "")
}
\arguments{
\item{object}{Stock-and-flow model, object of class \code{\link{sdbuildR}}.}

\item{name}{Name of unit. A character vector.}

\item{eqn}{Definition of unit. String or vector of unit definitions. Defaults to \code{1} to indicate a base unit not defined in terms of other units.}

\item{doc}{Documentation of unit.}
}
\value{
A stock-and-flow model object of class \code{\link{sdbuildR}}.
}
\description{
Add or change custom units in a stock-and-flow model. Custom units offer greater flexibility in defining units that are not part of the standard library. Custom units may be new base units, or may be defined in terms of other (custom) units. See \code{\link[=u]{u()}} for more information on the rules of specifying units. Note that units are only supported in Julia, not in R.
}
\examples{
\dontshow{if (Sys.getenv("NOT_CRAN") == "true" && is_julia_ready()) withAutoprint(\{ # examplesIf}
# Units are only supported with Julia
use_julia()
sfm <- sdbuildR("Crielaard2022")
sfm <- custom_unit(sfm, BMI, eqn = kg / m^2, doc = "Body Mass Index")

# You may also use words rather than symbols for the unit definition.
# The following modifies the unit BMI:
sfm <- custom_unit(sfm, BMI, eqn = kilogram / meters^2)

# Rename unit
sfm <- change_name(sfm, BMI, BodyMassIndex)

# Remove unit
sfm <- discard(sfm, BodyMassIndex)

# Unit names may need to be changed to be syntactically valid or to avoid
# overlap with existing units:
sfm <- custom_unit(sdbuildR(), C0^2)

# Close Julia session
use_julia(stop = TRUE)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=unit_prefixes]{unit_prefixes()}}, \code{\link[=discard]{discard()}}, \code{\link[=change_name]{change_name()}}
}
\concept{units}
