% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay.R
\name{delay}
\alias{delay}
\title{Retrieve a single value from the past}
\usage{
delay(variable, length, default_value = NULL)
}
\arguments{
\item{variable}{Variable of which to access past values.}

\item{length}{Length of the delay.}

\item{default_value}{Value to return when the delay length has not passed yet. If NULL, the value at the first time point is returned. Defaults to NULL.}
}
\value{
Past value
}
\description{
Also known as a pipeline or fixed delay. Note that this function is only usable for simulations run with Julia.
}
