% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/past.R
\name{delayN}
\alias{delayN}
\title{Delay of order N}
\usage{
delayN(variable, length, order, default_value = NULL)
}
\arguments{
\item{variable}{Variable to delay.}

\item{length}{Length of the delay.}

\item{order}{Order of the delay.}

\item{default_value}{Value to initialize the accumulator with. If NULL, the first value of the variable to delay is used. Defaults to NULL.}
}
\value{
Delayed variable
}
\description{
This function is coming soon, but is now unusable.
}
