% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_equations.R
\name{dependencies}
\alias{dependencies}
\title{Find dependencies}
\usage{
dependencies(object, name = NULL, type = NULL, reverse = FALSE)
}
\arguments{
\item{object}{Stock-and-flow model, object of class \code{\link{stockflow}}.}

\item{name}{Variable names to find dependencies for. Defaults to \code{NULL} to include all variables.}

\item{type}{Variable types to find dependencies for. Must be one or more of 'stock', 'flow', 'constant', 'aux', 'lookup', or 'func'. Defaults to \code{NULL} to include all types.}

\item{reverse}{If FALSE, list for each variable X which variables Y it depends on for its equation definition. If TRUE, don't show dependencies but dependents. This reverses the dependencies, such that for each variable X, it lists what other variables Y depend on X.}
}
\value{
List, with for each model variable what other variables it depends on, or if \code{reverse = TRUE}, which variables depend on it
}
\description{
Find which other variables each variable is dependent on.
}
\examples{
sfm <- stockflow("sir")
dependencies(sfm)

}
\concept{build}
