% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdbuildR_custom_func.R
\name{expit}
\alias{expit}
\title{Expit function}
\usage{
expit(x)
}
\arguments{
\item{x}{Numerical value}
}
\value{
Numerical value
}
\description{
Inverse of the logit function
}
\examples{
expit(1)
}
\concept{convenience}
