% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{filter_IM}
\alias{filter_IM}
\title{Filter vector based on logical function}
\usage{
filter_IM(y, condition_func)
}
\arguments{
\item{y}{Named vector}

\item{condition_func}{Function which takes arguments x and key and outputs a boolean value}
}
\value{
Vector with elements which meet condition_func
}
\description{
Equivalent of .Filter() in Insight Maker when using arguments x and key in the filter function
}
\examples{
filter_IM(c(a = 1, b = 2, c = 3), function(x, key) {
  x > 1
})
# b c
# 2 3
filter_IM(c(a = 1, b = 2, c = 3), function(x, key) {
  x > 1 & key != "b"
})
# c
# 3
}
\seealso{
Other custom: 
\code{\link{\%REM\%}()},
\code{\link{contains_IM}()},
\code{\link{expit}()},
\code{\link{indexof}()},
\code{\link{length_IM}()},
\code{\link{logistic}()},
\code{\link{logit}()},
\code{\link{mod}()},
\code{\link{rbool}()},
\code{\link{rdist}()},
\code{\link{rem}()},
\code{\link{round_IM}()},
\code{\link{shuffle}()},
\code{\link{substr_i}()}
}
\concept{custom}
