% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{find_dependencies}
\alias{find_dependencies}
\title{Find dependencies}
\usage{
find_dependencies(sfm)
}
\arguments{
\item{sfm}{Stock-and-flow model, object of class sdbuildR_xmile.}
}
\value{
Vector of dependencies (variable names in equation)
}
\description{
Find which other variables each variable is dependent on.
}
\examples{
sfm = xmile("SIR")
find_dependencies(sfm)

}
\seealso{
Other build: 
\code{\link{as.data.frame.sdbuildR_sim}()},
\code{\link{as.data.frame.sdbuildR_xmile}()},
\code{\link{build}()},
\code{\link{debugger}()},
\code{\link{get_build_code}()},
\code{\link{header}()},
\code{\link{macro}()},
\code{\link{plot.sdbuildR_xmile}()},
\code{\link{summary.sdbuildR_xmile}()},
\code{\link{xmile}()}
}
\concept{build}
