% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{get_build_code}
\alias{get_build_code}
\title{Generate code to build stock-and-flow model}
\usage{
get_build_code(object)
}
\arguments{
\item{object}{Stock-and-flow model, object of class \code{\link{sdbuildR}}.}
}
\value{
String with code to build stock-and-flow model from scratch.
}
\description{
Create R code to rebuild an existing stock-and-flow model. This may help to understand how a model is built, or to modify an existing one.
}
\examples{
sfm <- sdbuildR("SIR")
cat(get_build_code(sfm))

}
\concept{build}
