% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{get_logical_constraints}
\alias{get_logical_constraints}
\title{Generate constraints to use in ODE}
\usage{
get_logical_constraints(constraints_def)
}
\arguments{
\item{constraints_def}{Named list of minimum and maximum constraints per variable}
}
\value{
Vector of strings, specifying the statements to evaluate
}
\description{
Turn constraints on the minimum and maximum value a variable can take into a logical statement to evaluate in the ODE. Constraint variables can be stocks, flows, or any other variable defined in the ODE.
}
\examples{
constraints_def = list("Moose" = c(min = 0, max = 300), "Wolves" = c(max = 300))
get_logical_constraints(constraints_def) # c("Moose < 0", "Moose > 300", "Wolves > 300")
}
