% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{get_regex_time_units}
\alias{get_regex_time_units}
\title{Get regular expressions for time units in Julia}
\usage{
get_regex_time_units()
}
\value{
Named vector with regular expressions as names and units as entries
}
\description{
Get regular expressions for time units in Julia
}
\examples{
x <- get_regex_time_units()
head(x)

}
\concept{units}
