% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{get_regex_units}
\alias{get_regex_units}
\title{Get regular expressions for units in Julia}
\usage{
get_regex_units(object = NULL)
}
\arguments{
\item{object}{Stock-and-flow model, object of class \code{\link{sdbuildR}}.}
}
\value{
Named vector with regular expressions as names and units as entries
}
\description{
Get regular expressions for units in Julia
}
\examples{
x <- get_regex_units()
head(x)

}
\concept{units}
