% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{get_units}
\alias{get_units}
\title{View all standard units}
\usage{
get_units()
}
\value{
Dataframe with units in Julia
}
\description{
\code{get_units()} yields a dataframe with all standard units in Julia's Unitful package and added custom units by sdbuildR.
}
