% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{has_internet}
\alias{has_internet}
\title{Check if user has internet}
\usage{
has_internet()
}
\value{
Logical value
}
\description{
Internal function
}
\examples{
\dontshow{if (has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
has_internet()
\dontshow{\}) # examplesIf}
}
\seealso{
Other internal: 
\code{\link{compare_sim}()},
\code{\link{nonnegative}()},
\code{\link{not_on_cran}()},
\code{\link{saveat_func}()}
}
\concept{internal}
