% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{head.simulate_sdbuildR}
\alias{head.simulate_sdbuildR}
\title{Print first rows of a simulation}
\usage{
\method{head}{simulate_sdbuildR}(x, n = 6L, ...)
}
\arguments{
\item{x}{Output of \code{\link[=simulate.sdbuildR]{simulate()}}.}

\item{n}{Number of rows to print. Defaults to 6.}

\item{...}{Other arguments passed to \code{\link[=as.data.frame.simulate_sdbuildR]{as.data.frame.simulate_sdbuildR()}}.}
}
\value{
A data.frame with the first rows of the simulation results.
}
\description{
Print the first rows of a simulation data frame of a stock-and-flow model. This is a wrapper around \code{\link[=head]{head()}} that first converts the simulation results to a data frame using \link[=as.data.frame.simulate_sdbuildR]{as.data.frame()}.
}
\examples{
sfm <- sdbuildR("SIR")
sim <- simulate(sfm)
head(sim)
}
