% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{head.simulate_stockflow}
\alias{head.simulate_stockflow}
\title{Print first rows of a simulation}
\usage{
\method{head}{simulate_stockflow}(x, n = 6L, ...)
}
\arguments{
\item{x}{Output of \code{\link[=simulate.stockflow]{simulate()}}.}

\item{n}{Number of rows to print. Defaults to 6.}

\item{...}{Other arguments passed to \code{\link[=as.data.frame.simulate_stockflow]{as.data.frame.simulate_stockflow()}}.}
}
\value{
A data.frame with the first rows of the simulation results.
}
\description{
Print the first rows of a simulation data frame of a stock-and-flow model. This is a wrapper around \code{\link[=head]{head()}} that first converts the simulation results to a data frame using \link[=as.data.frame.simulate_stockflow]{as.data.frame()}.
}
\examples{
sfm <- stockflow("sir")
sim <- simulate(sfm)
head(sim)
}
\concept{simulate}
