% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{head.verify_stockflow}
\alias{head.verify_stockflow}
\title{Print first rows of verify results}
\usage{
\method{head}{verify_stockflow}(x, n = 6L, ...)
}
\arguments{
\item{x}{A \code{verify_stockflow} object.}

\item{n}{Number of rows. Defaults to 6.}

\item{...}{Other arguments passed to \code{\link[=as.data.frame.verify_stockflow]{as.data.frame.verify_stockflow()}}.}
}
\value{
A \code{data.frame}.
}
\description{
Wrapper around \code{\link[=head]{head()}} that first converts the results to a data frame using
\code{\link[=as.data.frame.verify_stockflow]{as.data.frame.verify_stockflow()}}.
}
\examples{
sfm <- stockflow("SIR") |>
  unit_test(expr = all(susceptible >= 0))
res <- verify(sfm)
head(res)
}
\concept{unitTest}
