% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_xmile.R
\name{header}
\alias{header}
\title{Modify header of stock-and-flow model}
\usage{
header(
  sfm,
  name = "My Model",
  caption = "My Model Description",
  created = Sys.time(),
  author = "Me",
  version = "1.0",
  URL = "",
  doi = "",
  ...
)
}
\arguments{
\item{sfm}{Stock-and-flow model, object of class \code{\link[=xmile]{sdbuildR_xmile}}.}

\item{name}{Model name. Defaults to "My Model".}

\item{caption}{Model description. Defaults to "My Model Description".}

\item{created}{Date the model was created. Defaults to Sys.time().}

\item{author}{Creator of the model. Defaults to "Me".}

\item{version}{Model version. Defaults to "1.0".}

\item{URL}{URL associated with model. Defaults to "".}

\item{doi}{DOI associated with the model. Defaults to "".}

\item{...}{Optional other entries to add to the header.}
}
\value{
A stock-and-flow model object of class \code{\link[=xmile]{sdbuildR_xmile}}
}
\description{
The header of a stock-and-flow model contains metadata about the model, such as the name, author, and version. Modify the header of an existing model with standard or custom properties.
}
\examples{
sfm <- xmile() |>
  header(
    name = "My first model",
    caption = "This is my first model",
    author = "Kyra Evers",
    version = "1.1"
  )
}
\concept{build}
