% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{IM_contains}
\alias{IM_contains}
\title{Check whether needle (value) is in haystack (vector or string)}
\usage{
IM_contains(haystack, needle)
}
\arguments{
\item{haystack}{Vector or string to search through}

\item{needle}{Value to search for}
}
\value{
Boolean
}
\description{
Equivalent of .Contains() in Insight Maker
}
\examples{
IM_contains(c("a", "b", "c"), "d") # FALSE
IM_contains(c("abcdef"), "bc") # TRUE
}
