% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{IM_filter}
\alias{IM_filter}
\title{Filter vector based on logical function}
\usage{
IM_filter(y, condition_func)
}
\arguments{
\item{y}{Named vector}

\item{condition_func}{Function which takes arguments x and key and outputs a boolean value}
}
\value{
Vector with elements which meet condition_func
}
\description{
Equivalent of .Filter() in Insight Maker when using arguments x and key in the filter function
}
\examples{
IM_filter(c(a = 1, b = 2, c = 3), function(x, key){x > 1})
# b c
# 2 3
IM_filter(c(a = 1, b = 2, c = 3), function(x, key){x > 1 & key != "b"})
# c
# 3
}
