% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{IM_length}
\alias{IM_length}
\title{Length of vector or string}
\usage{
IM_length(x)
}
\arguments{
\item{x}{A vector or a string}
}
\value{
The number of elements in x if x is a vector; the number of characters in x if x is a string
}
\description{
Equivalent of .Length() in Insight Maker, which returns the number of elements when performed on a vector, but returns the number of characters when performed on a string
}
\examples{
IM_length(c("a", "b", "c")) # 3
IM_length("abcdef") # 6
IM_length(c("abcdef")) # 6
}
