% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_unit_test.R
\name{interpret}
\alias{interpret}
\title{Recursively interpret a parsed R expression}
\usage{
interpret(e, parent_op = NULL)
}
\arguments{
\item{e}{A language object (from parse())}

\item{parent_op}{The operator of the parent call (used for precedence decisions)}
}
\value{
A human-readable string
}
\description{
Recursively interpret a parsed R expression
}
\keyword{internal}
