% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_julia.R
\name{is_julia_ready}
\alias{is_julia_ready}
\title{Check if Julia is ready to be used with sdbuildR}
\usage{
is_julia_ready()
}
\value{
Logical value. \code{TRUE} if Julia is ready to be used with sdbuildR, \code{FALSE} otherwise.
}
\description{
Check if Julia can be started and if the Julia environment for sdbuildR has been instantiated.
}
\examples{
is_julia_ready()
}
\seealso{
\code{\link[=install_julia_env]{install_julia_env()}}, \code{\link[=use_julia]{use_julia()}}
}
