% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_julia.R
\name{julia_setup_ok}
\alias{julia_setup_ok}
\title{Check Julia environment set up}
\usage{
julia_setup_ok()
}
\value{
Logical value
}
\description{
Check Julia environment set up
}
\examples{
julia_setup_ok()
}
\seealso{
Other simulate: 
\code{\link{ensemble}()},
\code{\link{plot.sdbuildR_ensemble}()},
\code{\link{plot.sdbuildR_sim}()},
\code{\link{sim_specs}()},
\code{\link{simulate}()},
\code{\link{use_julia}()},
\code{\link{use_threads}()}
}
\concept{simulate}
