% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_setup.R
\name{julia_status}
\alias{julia_status}
\title{Check status of Julia installation and environment}
\usage{
julia_status(verbose = TRUE)
}
\arguments{
\item{verbose}{If TRUE, print detailed status information. Defaults to TRUE.}
}
\value{
A list with components:
\item{julia_found}{Logical. TRUE if Julia installation found.}
\item{julia_path}{Character. Path to Julia bin.}
\item{julia_version}{Character. Julia version string, or NULL if not found.}
\item{env_exists}{Logical. TRUE if inst/Project.toml exists in sdbuildR package.}
\item{env_instantiated}{Logical. TRUE if inst/Manifest.toml exists (i.e., Julia environment was instantiated).}
\item{status}{Character. Overall status: "julia_not_installed", "julia_needs_update", "sdbuildR_needs_reinstall", "install_julia_env", "ready", or "unknown".}
}
\description{
Check if Julia can be found and if the Julia environment for sdbuildR has been instantiated. Note that this does not mean a Julia session has been started, merely whether it \emph{could} be. For more guidance, please see \href{https://kcevers.github.io/sdbuildR/articles/julia-setup.html}{this vignette}.
}
\examples{
status <- julia_status()
print(status)

}
\seealso{
Other julia: 
\code{\link{install_julia_env}()},
\code{\link{use_julia}()}
}
\concept{julia}
