% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdbuildR_custom_func.R
\name{length_IM}
\alias{length_IM}
\title{Length of vector or string}
\usage{
length_IM(x)
}
\arguments{
\item{x}{A vector or a string}
}
\value{
The number of elements in x if x is a vector; the number of characters in x if x is a string
}
\description{
Equivalent of .Length() in Insight Maker, which returns the number of elements when performed on a vector, but returns the number of characters when performed on a string
}
\examples{
length_IM(c("a", "b", "c")) # 3
length_IM("abcdef") # 6
}
\concept{custom}
