% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdbuildR_custom_func.R
\name{logit}
\alias{logit}
\title{Logit function}
\usage{
logit(p)
}
\arguments{
\item{p}{Probability, numerical value between 0 and 1}
}
\value{
Numerical value
}
\description{
Logit function
}
\examples{
logit(.1)
}
\concept{convenience}
