% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_unit_test.R
\name{maybe_wrap}
\alias{maybe_wrap}
\title{Conditionally wrap a result string in parentheses}
\usage{
maybe_wrap(result, current_op, parent_op)
}
\arguments{
\item{result}{The human-readable string for this sub-expression}

\item{current_op}{The current operator}

\item{parent_op}{The parent operator (from the recursive call)}
}
\value{
Possibly parenthesized string
}
\description{
Used by infix operators to add parens when the current operator
has lower precedence than the parent context.
}
\keyword{internal}
