% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{meta}
\alias{meta}
\title{Modify meta of stock-and-flow model}
\usage{
meta(
  object,
  name = "My Model",
  caption = "My Model Description",
  created = Sys.time(),
  author = "Me",
  version = "0.0.0.9000",
  URL = "",
  doi = "",
  ...
)
}
\arguments{
\item{object}{Stock-and-flow model, object of class \code{\link{stockflow}}.}

\item{name}{Model name. Defaults to "My Model".}

\item{caption}{Model description. Defaults to "My Model Description".}

\item{created}{Date the model was created. Defaults to Sys.time().}

\item{author}{Creator of the model. Defaults to "Me".}

\item{version}{Model version. Defaults to "0.0.0.9000", a development version.}

\item{URL}{URL associated with model. Defaults to "".}

\item{doi}{DOI associated with the model. Defaults to "".}

\item{...}{Optional other entries to add to the meta.}
}
\value{
A stock-and-flow model object of class \code{\link{stockflow}}
}
\description{
The meta of a stock-and-flow model contains metadata about the model, such as the name, author, and version. Modify the meta of an existing model with standard or custom properties.
}
\examples{
sfm <- stockflow() |>
  meta(
    name = "My first model",
    caption = "This is my first model",
    author = "Kyra Evers",
    version = "1.1"
  )
}
\concept{build}
