% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{mod}
\alias{mod}
\title{Modulus}
\usage{
mod(a, b)
}
\arguments{
\item{a}{Dividend}

\item{b}{Divisor}
}
\value{
Modulus
}
\description{
Note that modulus and remainder are not the same in case either a or b is negative. If you work with negative numbers, modulus is always non-negative (it matches the sign of the divisor).
}
\examples{

# mod(a, b) is the same as a \%\% b

a = 7; b = 3
rem(a, b)
mod(a, b)
a = -7; b = 3
rem(a, b)
mod(a, b)
a = 7; b = -3
rem(a, b)
mod(a, b)
a = -7; b = -3
rem(a, b)
mod(a, b)
}
