% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdbuildR_custom_func.R
\name{nonnegative}
\alias{nonnegative}
\title{Check whether x is less than zero}
\usage{
nonnegative(x)
}
\arguments{
\item{x}{Value}
}
\value{
x if x is greater than 0, 0 otherwise
}
\description{
Check whether x is less than zero.
}
\examples{
nonnegative(NA)
nonnegative(-1)

}
\concept{internal}
