% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{not_on_cran}
\alias{not_on_cran}
\title{Check if on CRAN}
\usage{
not_on_cran()
}
\value{
Logical value
}
\description{
Check if on CRAN
}
\examples{
not_on_cran()
}
\seealso{
Other internal: 
\code{\link{compare_sim}()},
\code{\link{has_internet}()},
\code{\link{nonnegative}()},
\code{\link{saveat_func}()}
}
\concept{internal}
