% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay.R
\name{past}
\alias{past}
\title{Retrieve values from the past}
\usage{
past(variable, length = NULL)
}
\arguments{
\item{variable}{Variable of which to access past values.}

\item{length}{Interval from the current time point to look back to. If NULL, all available past values are returned. Defaults to NULL.}
}
\value{
Past values
}
\description{
Function in development; currently unusable. Note that this function is only usable for simulations run with Julia.
}
\seealso{
Other delay: 
\code{\link{delay}()},
\code{\link{delayN}()},
\code{\link{smoothN}()}
}
\concept{delay}
