% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_xmile.R
\name{plot.sdbuildR_ensemble}
\alias{plot.sdbuildR_ensemble}
\title{Plot timeseries of ensemble}
\usage{
\method{plot}{sdbuildR_ensemble}(
  x,
  i = 1:10,
  j = 1,
  palette = "Dark 2",
  colors = NULL,
  font_family = "Times New Roman",
  font_size = 16,
  ...
)
}
\arguments{
\item{x}{Output of ensemble().}

\item{i}{Indices of the individual trajectories to plot if type = "sims". Defaults to 1:10. Including a high number of trajectories will slow down plotting considerably.}

\item{j}{Index of the condition to plot. Defaults to 1.}

\item{palette}{Colour palette. Must be one of hcl.pals().}

\item{colors}{Vector of colours. If NULL, the colour palette will be used. If specified, will override palette. The number of colours must be equal to the number of variables in the simulation dataframe. Defaults to NULL.}

\item{font_family}{Font family. Defaults to "Times New Roman".}

\item{font_size}{Font size. Defaults to 16.}

\item{...}{Optional parameters}
}
\value{
Plot of ensemble simulation.
}
\description{
Visualise ensemble simulation results of a stock-and-flow model. Plot the evolution of stocks over time, with the option of also showing other model variables.
}
\seealso{
\code{\link[=ensemble]{ensemble()}}
}
