% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_xmile.R
\name{plot.sdbuildR_sim}
\alias{plot.sdbuildR_sim}
\title{Plot timeseries of simulation}
\usage{
\method{plot}{sdbuildR_sim}(
  x,
  add_constants = FALSE,
  palette = "Dark 2",
  colors = NULL,
  title = x[["sfm"]][["header"]][["name"]],
  font_family = "Times New Roman",
  font_size = 16,
  ...
)
}
\arguments{
\item{x}{Output of simulate().}

\item{add_constants}{If TRUE, include constants in plot. Defaults to FALSE.}

\item{palette}{Colour palette. Must be one of hcl.pals().}

\item{colors}{Vector of colours. If NULL, the colour palette will be used. If specified, will override palette. The number of colours must be equal to the number of variables in the simulation dataframe. Defaults to NULL.}

\item{title}{Plot title. Character. Defaults to the name in the header of the model.}

\item{font_family}{Font family. Defaults to "Times New Roman".}

\item{font_size}{Font size. Defaults to 16.}

\item{...}{Optional parameters}
}
\value{
Plot object
}
\description{
Visualise simulation results of a stock-and-flow model. Plot the evolution of stocks over time, with the option of also showing other model variables.
}
\examples{
sfm = xmile("SIR")
sim = simulate(sfm)
plot(sim)
}
\seealso{
\code{\link[=simulate]{simulate()}}
}
