% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plot_sdbuildR2}
\alias{plot_sdbuildR2}
\title{Plot stock-and-flow diagram with qgraph}
\usage{
plot_sdbuildR2(
  x,
  vars = NULL,
  format_label = TRUE,
  wrap_width = 20,
  font_size = 18,
  font_family = "Times New Roman",
  stock_col = "#83d3d4",
  flow_col = "#f48153",
  dependency_col = "#999999",
  show_dependencies = TRUE,
  show_constants = FALSE,
  show_aux = TRUE,
  layout_algo = c("sugiyama", "fr"),
  level_spacing = 1.25,
  node_spacing = 1,
  show_eqn = FALSE,
  ...
)
}
\arguments{
\item{x}{A stock-and-flow model object of class \code{\link{sdbuildR}}.}

\item{vars}{Variables to plot. Defaults to NULL to plot all variables.}

\item{format_label}{If TRUE, apply default formatting (removing periods and underscores) to labels if labels are the same as variable names.}

\item{wrap_width}{Width of text wrapping for labels. Must be an integer. Defaults to 20.}

\item{font_size}{Font size. Defaults to 18.}

\item{font_family}{Font name. Defaults to "Times New Roman".}

\item{stock_col}{Colour of stocks. Defaults to "#83d3d4".}

\item{flow_col}{Colour of flows. Defaults to "#f48153".}

\item{dependency_col}{Colour of dependency arrows. Defaults to "#999999".}

\item{show_dependencies}{If TRUE, show dependencies between variables. Defaults to TRUE.}

\item{show_constants}{If TRUE, show constants. Defaults to FALSE.}

\item{show_aux}{If TRUE, show auxiliary variables. Defaults to TRUE.}

\item{layout_algo}{Layout algorithm. One of "sugiyama" (structured, default) or "fr".}

\item{level_spacing}{Horizontal spacing multiplier between layout levels.}

\item{node_spacing}{Vertical spacing multiplier between nodes.}

\item{show_eqn}{If TRUE, attach equations as metadata attribute on the returned object.}

\item{...}{Optional arguments reserved for future extensions.}
}
\value{
qgraph object.
}
\description{
Visualize a stock-and-flow diagram using the R package qgraph. This function
is a static-graphics alternative to \code{\link[=plot.sdbuildR]{plot.sdbuildR()}}, preserving core
filtering and labeling behavior while using a deterministic hierarchical
layout.
}
\examples{
\dontrun{
sfm <- sdbuildR("SIR")
pl <- plot_sdbuildR2(sfm)
export_plot(pl, tempfile(fileext = ".pdf"))
}
}
\seealso{
\code{\link[=plot.sdbuildR]{plot.sdbuildR()}}, \code{\link[=export_plot]{export_plot()}}
}
\concept{build}
