% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{print.compare_sdbuildR}
\alias{print.compare_sdbuildR}
\title{Print comparison of two stock-and-flow models}
\usage{
\method{print}{compare_sdbuildR}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link[=compare_models]{compare_sdbuildR}}}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print comparison of two stock-and-flow models
}
\concept{build}
