% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{print.diagnose_sdbuildR}
\alias{print.diagnose_sdbuildR}
\title{Print method for diagnose_sdbuildR}
\usage{
\method{print}{diagnose_sdbuildR}(x, ...)
}
\arguments{
\item{x}{Object of class \code{diagnose_sdbuildR}}

\item{...}{Ignored}
}
\value{
\code{x} invisibly
}
\description{
Print method for diagnose_sdbuildR
}
