% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{print.simulate_sdbuildR}
\alias{print.simulate_sdbuildR}
\title{Print simulation of a stock-and-flow model}
\usage{
\method{print}{simulate_sdbuildR}(x, ...)
}
\arguments{
\item{x}{A simulation result of class \code{\link[=simulate.sdbuildR]{simulate_sdbuildR}}}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns \code{x}
}
\description{
Prints the first rows of the simulation results in wide format. For a
statistical summary per variable use \link[=summary.simulate_sdbuildR]{summary()}.
}
\examples{
sfm <- sdbuildR("SIR")
sim <- simulate(sfm)
print(sim)

}
\seealso{
\code{\link[=simulate.sdbuildR]{simulate.sdbuildR()}}, \code{\link[=summary.simulate_sdbuildR]{summary.simulate_sdbuildR()}},
\code{\link[=plot.simulate_sdbuildR]{plot.simulate_sdbuildR()}}, \code{\link[=as.data.frame.simulate_sdbuildR]{as.data.frame.simulate_sdbuildR()}}
}
\concept{build}
