% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{print.simulate_stockflow}
\alias{print.simulate_stockflow}
\title{Print simulation of a stock-and-flow model}
\usage{
\method{print}{simulate_stockflow}(x, ...)
}
\arguments{
\item{x}{A simulation result of class \code{\link[=simulate.stockflow]{simulate_stockflow}}}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns \code{x}
}
\description{
Prints the first rows of the simulation results in wide format. For a
statistical summary per variable use \link[=summary.simulate_stockflow]{summary()}.
}
\examples{
sfm <- stockflow("SIR")
sim <- simulate(sfm)
print(sim)

}
\seealso{
\code{\link[=simulate.stockflow]{simulate.stockflow()}}, \code{\link[=summary.simulate_stockflow]{summary.simulate_stockflow()}},
\code{\link[=plot.simulate_stockflow]{plot.simulate_stockflow()}}, \code{\link[=as.data.frame.simulate_stockflow]{as.data.frame.simulate_stockflow()}}
}
\concept{build}
