% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_xmile.R
\name{print.summary.sdbuildR_xmile}
\alias{print.summary.sdbuildR_xmile}
\title{Print method for summary.sdbuildR_xmile}
\usage{
\method{print}{summary.sdbuildR_xmile}(x, ...)
}
\arguments{
\item{x}{A summary object of class \link{summary.sdbuildR_xmile}}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns the summary object of class \link{summary.sdbuildR_xmile}
}
\description{
Print method for summary.sdbuildR_xmile
}
\concept{build}
