% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdbuildR.R
\name{print.summary_sdbuildR}
\alias{print.summary_sdbuildR}
\title{Print summary of stock-and-flow model}
\usage{
\method{print}{summary_sdbuildR}(x, ...)
}
\arguments{
\item{x}{A summary object of class \code{\link[=summary.sdbuildR]{summary_sdbuildR}}}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns the summary object of class \code{\link[=summary.sdbuildR]{summary_sdbuildR}}
}
\description{
Print summary of stock-and-flow model
}
\concept{build}
