% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stockflow.R
\name{print.summary_stockflow}
\alias{print.summary_stockflow}
\title{Print method for summary_stockflow}
\usage{
\method{print}{summary_stockflow}(x, ...)
}
\arguments{
\item{x}{Object of class \code{summary_stockflow}}

\item{...}{Ignored}
}
\value{
\code{x} invisibly
}
\description{
Print method for summary_stockflow
}
\concept{build}
