% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdbuildR_custom_func.R
\name{rbool}
\alias{rbool}
\title{Generate random logical value}
\usage{
rbool(p)
}
\arguments{
\item{p}{Probability of TRUE, numerical value between 0 and 1}
}
\value{
Logical value
}
\description{
Equivalent of RandBoolean() in Insight Maker
}
\examples{
rbool(.5)
}
\concept{convenience}
