% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{rem}
\alias{rem}
\title{Remainder}
\usage{
rem(a, b)
}
\arguments{
\item{a}{Dividend}

\item{b}{Divisor}
}
\value{
Remainder
}
\description{
Note that modulus and remainder are not the same in case either a or b is negative. If you work with negative numbers, modulus is always non-negative (it matches the sign of the divisor).
}
\examples{
a = 7; b = 3
rem(a, b)
mod(a, b)
a = -7; b = 3
rem(a, b)
mod(a, b)
a = 7; b = -3
rem(a, b)
mod(a, b)
a = -7; b = -3
rem(a, b)
mod(a, b)
}
