% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdbuildR-methods.R
\name{sdbuildR-as.data.frame}
\alias{sdbuildR-as.data.frame}
\alias{as.data.frame}
\title{Data frame methods for sdbuildR objects}
\description{
\code{as.data.frame()} extracts the contents of an sdbuildR object as a data
frame:
}
\details{
\describe{
\item{\code{\link[=as.data.frame.stockflow]{as.data.frame.stockflow()}}}{Structure of a model: its stocks, flows, constants and auxiliaries.}
\item{\code{\link[=as.data.frame.simulate_stockflow]{as.data.frame.simulate_stockflow()}}}{Results of a simulation.}
\item{\code{\link[=as.data.frame.ensemble_stockflow]{as.data.frame.ensemble_stockflow()}}}{Results of an ensemble simulation.}
\item{\code{\link[=as.data.frame.verify_stockflow]{as.data.frame.verify_stockflow()}}}{Results of \code{\link[=verify]{verify()}}.}
}
}
\seealso{
\link[=sdbuildR-head-tail]{head()} and \link[=sdbuildR-head-tail]{tail()} methods
}
\concept{methods}
