% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdbuildR-methods.R
\name{sdbuildR-head-tail}
\alias{sdbuildR-head-tail}
\alias{head}
\alias{tail}
\title{First and last rows of sdbuildR results}
\description{
\code{head()} and \code{tail()} return the first or last rows of the data frame
produced by \code{as.data.frame()}:
}
\details{
\describe{
\item{\code{\link[=head.simulate_stockflow]{head.simulate_stockflow()}}, \code{\link[=tail.simulate_stockflow]{tail.simulate_stockflow()}}}{Rows of a simulation.}
\item{\code{\link[=head.verify_stockflow]{head.verify_stockflow()}}, \code{\link[=tail.verify_stockflow]{tail.verify_stockflow()}}}{Rows of \code{\link[=verify]{verify()}} results.}
}

\code{head()} and \code{tail()} methods are also defined for
\code{\link[=ensemble]{ensemble_stockflow}} objects.
}
\seealso{
\link[=sdbuildR-as.data.frame]{as.data.frame()} methods
}
\concept{methods}
