% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdbuildR-methods.R
\name{sdbuildR-plot}
\alias{sdbuildR-plot}
\alias{plot}
\title{Plot methods for sdbuildR objects}
\description{
sdbuildR provides a \code{plot()} method for each of its object classes:
}
\details{
\describe{
\item{\code{\link[=plot.stockflow]{plot.stockflow()}}}{Stock-and-flow diagram of a model.}
\item{\code{\link[=plot.simulate_stockflow]{plot.simulate_stockflow()}}}{Timeseries of a simulation.}
\item{\code{\link[=plot.ensemble_stockflow]{plot.ensemble_stockflow()}}}{Timeseries of an ensemble simulation.}
\item{\code{\link[=plot.verify_stockflow]{plot.verify_stockflow()}}}{Results of \code{\link[=verify]{verify()}}.}
}
}
\seealso{
\code{\link[=stockflow]{stockflow()}}, \code{\link[=simulate.stockflow]{simulate()}}, \code{\link[=ensemble]{ensemble()}}, \code{\link[=verify]{verify()}}
}
\concept{methods}
