% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_julia.R
\name{sdbuildR_setup}
\alias{sdbuildR_setup}
\title{Set up Julia environment for sdbuildR}
\usage{
sdbuildR_setup(
  version = "latest",
  JULIA_HOME = NULL,
  dir = NULL,
  force = FALSE,
  ...
)
}
\arguments{
\item{version}{Julia version. Default is "latest", which will install the most recent stable release.}

\item{JULIA_HOME}{Path to Julia installation. Defaults to NULL to locate Julia automatically.}

\item{dir}{Directory to install Julia. Defaults to NULL to find default location.}

\item{force}{If TRUE, force Julia installation even if existing version is found. Defaults to FALSE.}

\item{...}{Optional arguments passed to JuliaCall::julia_setup()}
}
\description{
Create Julia environment to simulate stock-and-flow models. \code{sdbuildR_setup()} looks for a Julia installation, and will install Julia as well as some required packages if not found. Keep in mind that the first time running \code{sdbuildR_setup()} can take around 5-15 minutes, and all subsequent calls can take around 1 minute.
}
\examples{
sdbuildR_setup()
}
