% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{shuffle}
\alias{shuffle}
\title{Shuffle vector}
\usage{
shuffle(x)
}
\arguments{
\item{x}{Vector to shuffle}
}
\value{
Shuffled x
}
\description{
Shuffle vector
}
\examples{
shuffle(1:10)
}
\seealso{
Other custom: 
\code{\link{\%REM\%}()},
\code{\link{contains_IM}()},
\code{\link{expit}()},
\code{\link{filter_IM}()},
\code{\link{indexof}()},
\code{\link{length_IM}()},
\code{\link{logistic}()},
\code{\link{logit}()},
\code{\link{mod}()},
\code{\link{rbool}()},
\code{\link{rdist}()},
\code{\link{rem}()},
\code{\link{round_IM}()},
\code{\link{substr_i}()}
}
\concept{custom}
