% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid function (i.e. logistic function)}
\usage{
sigmoid(x, slope = 1, midpoint = 0)
}
\arguments{
\item{x}{Value to compute sigmoid of}

\item{slope}{Slope of sigmoid function, defaults to 1}

\item{midpoint}{Midpoint of sigmoid function where the output is .5, defaults to 0}
}
\value{
Sigmoid of x
}
\description{
Sigmoid function (i.e. logistic function)
}
\examples{
sigmoid(0)
sigmoid(1, slope = 5, midpoint = 0.5)
}
