% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.sdbuildR}
\alias{simulate.sdbuildR}
\title{Simulate stock-and-flow model}
\usage{
\method{simulate}{sdbuildR}(object, nsim = 1, seed = NULL, verbose = FALSE, ...)
}
\arguments{
\item{object}{Stock-and-flow model, object of class \code{\link{sdbuildR}}.}

\item{nsim}{Number of simulations to run (currently unused; see \code{\link[=ensemble]{ensemble()}} for running multiple simulations).}

\item{seed}{Seed number to ensure reproducibility across runs in case of
random elements. Must be an integer. Defaults to \code{NULL} (no seed).}

\item{verbose}{If \code{TRUE}, print duration of simulation. Defaults to \code{FALSE}.}

\item{...}{Optional arguments passed to \code{\link[=sim_specs]{sim_specs()}}; these can be used to override the simulation specifications set in the model object.}
}
\value{
Object of class \code{\link[=simulate.sdbuildR]{simulate_sdbuildR}}, a list containing:
\describe{
\item{object}{Stock-and-flow model object of class \code{\link{sdbuildR}}}
\item{df}{Data frame: simulation results (time, variable, value)}
\item{init}{Named vector: initial stock values}
\item{constants}{Named vector: constant parameters}
\item{script}{Character: generated simulation code (R or Julia)}
\item{duration}{Numeric: simulation time in seconds}
\item{success}{Logical: TRUE if completed without errors}
\item{error_message}{NULL if completed without errors}
}

Use \link[=as.data.frame.simulate_sdbuildR]{as.data.frame()} to extract results, \link[=plot.simulate_sdbuildR]{plot()} to visualize.
}
\description{
Simulate a stock-and-flow model with simulation specifications defined by \code{\link[=sim_specs]{sim_specs()}}. If \code{sim_specs(language = "julia")}, the Julia environment will first be set up with \code{\link[=use_julia]{use_julia()}}. If any problems are detected by \code{\link[=diagnose]{diagnose()}}, the model cannot be simulated.
}
\examples{
sfm <- sdbuildR("SIR")
sim <- simulate(sfm)
plot(sim)

# Obtain all model variables
sim <- simulate(sim_specs(sfm, only_stocks = FALSE))
plot(sim, add_constants = TRUE)

\dontshow{if (Sys.getenv("NOT_CRAN") == "true" && is_julia_ready()) withAutoprint(\{ # examplesIf}
# Use Julia for models with units
sfm <- sim_specs(sdbuildR("coffee_cup"), language = "Julia")
sim <- simulate(sfm)
plot(sim)

# Close Julia session
use_julia(stop = TRUE)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=update]{update()}}, \code{\link[=sdbuildR]{sdbuildR()}}, \code{\link[=diagnose]{diagnose()}}, \code{\link[=sim_specs]{sim_specs()}}, \code{\link[=use_julia]{use_julia()}}
}
\concept{simulate}
