% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{tail.verify_sdbuildR}
\alias{tail.verify_sdbuildR}
\title{Print last rows of verify results}
\usage{
\method{tail}{verify_sdbuildR}(x, n = 6L, ...)
}
\arguments{
\item{x}{A \code{verify_sdbuildR} object.}

\item{n}{Number of rows. Defaults to 6.}

\item{...}{Other arguments passed to \code{\link[=as.data.frame.verify_sdbuildR]{as.data.frame.verify_sdbuildR()}}.}
}
\value{
A \code{data.frame}.
}
\description{
Wrapper around \code{\link[=tail]{tail()}} that first converts the results to a data frame using
\code{\link[=as.data.frame.verify_sdbuildR]{as.data.frame.verify_sdbuildR()}}.
}
\examples{
sfm <- sdbuildR("SIR") |>
  unit_test(expr = all(Susceptible >= 0))
res <- verify(sfm)
tail(res)
}
\concept{unitTest}
