% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{unit_tests}
\alias{unit_tests}
\title{Display unit tests defined on a stock-and-flow model}
\usage{
unit_tests(object)
}
\arguments{
\item{object}{Stock-and-flow model, object of class \code{\link{sdbuildR}}.}
}
\value{
An object of class \code{unit_tests_sdbuildR}, printed automatically.
}
\description{
Returns an overview of all unit tests attached to the model. The result
has a \code{print()} method.
}
\examples{
sfm <- sdbuildR("SIR") |>
  unit_test(expr = all(Susceptible >= 0)) |>
  unit_test(label = "Recovered increases over time", 
            expr = all(diff(Recovered) >= 0))

unit_tests(sfm)
}
\seealso{
\code{\link[=unit_test]{unit_test()}}, \code{\link[=verify]{verify()}}
}
\concept{verify}
