% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insightmaker_conv.R
\name{url_to_insightmaker}
\alias{url_to_insightmaker}
\title{Extract Insight Maker model from URL}
\usage{
url_to_insightmaker(url, file = NULL)
}
\arguments{
\item{url}{String with URL to an Insight Maker model}

\item{file}{If specified, file path to save Insight Maker model to. If NULL, do not save model.}
}
\value{
XML string with Insight Maker model
}
\description{
Create XML string from Insight Maker URL. This function is for internal use; use \code{import_insightmaker()} to import an Insight Maker model.
}
\examples{
\dontshow{if (has_internet()) withAutoprint(\{ # examplesIf}
url <- "https://insightmaker.com/insight/43tz1nvUgbIiIOGSGtzIzj/Romeo-Juliet"
xml <- url_to_insightmaker(url)

# Save model to file
file <- tempfile(fileext = ".InsightMaker")
xml <- url_to_insightmaker(url, file = file)
file.remove(file)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=import_insightmaker]{import_insightmaker()}}
}
\concept{importExport}
