% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{verify}
\alias{verify}
\title{Verify model}
\usage{
verify(object, ...)
}
\arguments{
\item{object}{A model object to verify.}

\item{...}{Additional arguments passed to specific methods.}
}
\description{
Verify model
}
