% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_xmile.R
\name{as.data.frame.sdbuildR_sim}
\alias{as.data.frame.sdbuildR_sim}
\title{Create dataframe of simulation results}
\usage{
\method{as.data.frame}{sdbuildR_sim}(x, ...)
}
\arguments{
\item{x}{Output of simulate().}

\item{...}{Optional parameters}
}
\value{
Dataframe with simulation results
}
\description{
Convert simulation results to a dataframe. The first column is time, followed by all stocks, and then all other auxiliary and flow variables.
}
\examples{
sfm = xmile("SIR")
sim = simulate(sfm)
head(as.data.frame(sim))


}
\seealso{
\code{\link[=simulate]{simulate()}}, \code{\link[=xmile]{xmile()}}
}
