% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_xmile.R
\name{as.data.frame.sdbuildR_sim}
\alias{as.data.frame.sdbuildR_sim}
\title{Create dataframe of simulation results}
\usage{
\method{as.data.frame}{sdbuildR_sim}(x, row.names = NULL, optional = FALSE, direction = "long", ...)
}
\arguments{
\item{x}{Output of simulate().}

\item{row.names}{NULL or a character vector giving the row names for the data frame. Missing values are not allowed.}

\item{optional}{Ignored parameter.}

\item{direction}{Format of dataframe, either "long" (default) or "wide".}

\item{...}{Optional parameters}
}
\value{
Dataframe with simulation results
}
\description{
Convert simulation results to a dataframe. The first column is time, followed by all stocks, and then all other auxiliary and flow variables.
}
\examples{
sfm = xmile("SIR")
sim = simulate(sfm)
head(as.data.frame(sim))


}
\seealso{
\code{\link[=simulate]{simulate()}}, \code{\link[=xmile]{xmile()}}

Other build: 
\code{\link{as.data.frame.sdbuildR_xmile}()},
\code{\link{build}()},
\code{\link{debugger}()},
\code{\link{find_dependencies}()},
\code{\link{get_build_code}()},
\code{\link{header}()},
\code{\link{macro}()},
\code{\link{plot.sdbuildR_xmile}()},
\code{\link{summary.sdbuildR_xmile}()},
\code{\link{xmile}()}
}
\concept{build}
