% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_xmile.R
\name{as.data.frame.sdbuildR_xmile}
\alias{as.data.frame.sdbuildR_xmile}
\title{Convert stock-and-flow model to dataframe}
\usage{
\method{as.data.frame}{sdbuildR_xmile}(
  x,
  row.names = NULL,
  optional = FALSE,
  type = NULL,
  name = NULL,
  properties = NULL,
  ...
)
}
\arguments{
\item{x}{Stock-and-flow model of class sdbuildR_xmile}

\item{row.names}{NULL or a character vector giving the row names for the data frame. Missing values are not allowed.}

\item{optional}{Ignored parameter.}

\item{type}{Variable types to retain in the dataframe. Must be one or more of 'stock', 'flow', 'constant', 'aux', 'gf', 'model_units', or 'macro'. Defaults to NULL to include all variable types.}

\item{name}{Variable names to retain in the dataframe. Defaults to NULL to include all variables.}

\item{properties}{Variable properties to retain in the dataframe. Defaults to NULL to include all properties.}

\item{...}{Optional arguments}
}
\value{
Dataframe with properties of all model variables, model units, and macros.
}
\description{
Create a dataframe with properties of all model variables, model units, and macros. Specify the variable types, variable names, and/or properties to get a subset of the dataframe.
}
\examples{
as.data.frame(xmile("SIR"))

# Only show stocks
as.data.frame(xmile("SIR"), type = "stock")

# Only show equation and label
as.data.frame(xmile("SIR"), properties = c("eqn", "label"))
}
\seealso{
Other build: 
\code{\link{as.data.frame.sdbuildR_sim}()},
\code{\link{build}()},
\code{\link{debugger}()},
\code{\link{get_build_code}()},
\code{\link{header}()},
\code{\link{macro}()},
\code{\link{plot.sdbuildR_xmile}()},
\code{\link{summary.sdbuildR_xmile}()},
\code{\link{xmile}()}
}
\concept{build}
