% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{change_type}
\alias{change_type}
\title{Change variable type}
\usage{
change_type(sfm, name, new_type)
}
\arguments{
\item{sfm}{Stock-and-flow model, object of class \code{\link{sdbuildR}}.}

\item{name}{Variable name. Accepts a bare symbol (e.g., \code{population}), a string (\code{"population"}), or a vector via \code{c()} (e.g., \code{c(a, b)} or \code{c("a", "b")}). Use \verb{!!} to inject from a variable.}

\item{new_type}{New variable type; one of 'stock', 'flow', 'constant', 'aux', 'gf', or 'func'. Character vector of the same length as name. If NULL, types will be validated but not changed.}
}
\value{
A stock-and-flow model object of class \code{\link[=sdbuildR]{sdbuildR}} with the variable type changed throughout the model. Note that changing the type may result in changes to other properties (e.g. a flow must have "to" and/or "from" properties, so these will be added if not already present), and may require changes to the equations of connected variables.
}
\description{
Change the type of a variable in a stock-and-flow model.
}
\examples{
# Change the birth rate of predators from a constant to an auxiliary
sfm <- sdbuildR("predator_prey")
sfm <- change_type(sfm, delta, new_type = aux) |>
       # Use a sin function to introduce seasonality in the birth rate
      build(delta, eqn = 0.025 + 0.01 * sin(2 * pi * t / 12))
sim <- simulate(sfm)
plot(sim)

}
\seealso{
\code{\link[=build]{build()}}
}
\concept{build}
