% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clean_name}
\alias{clean_name}
\title{Clean variable name(s)}
\usage{
clean_name(new, existing, protected = c())
}
\arguments{
\item{new}{Vector of names to transform to valid names}

\item{existing}{Vector of existing names in model}

\item{protected}{Optional vector of protected names}
}
\value{
Translated names
}
\description{
Clean variable name(s) to create syntactically valid, unique names for use in R and Julia.
}
\seealso{
Other internal: 
\code{\link{compare_sim}()},
\code{\link{has_internet}()},
\code{\link{nonnegative}()},
\code{\link{not_on_cran}()},
\code{\link{saveat_func}()}
}
\concept{internal}
