% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/past.R
\name{compute_smoothN}
\alias{compute_smoothN}
\title{Compute the smoothN accumulator}
\usage{
compute_smoothN(input, state, length, order)
}
\arguments{
\item{input}{Input value to process.}

\item{state}{Current state of the accumulator.}

\item{length}{Length of the smoothing.}

\item{order}{Order of the smoothing.}
}
\value{
A list containing the outflow value and the updated state.
}
\description{
Internal function to compute the smoothN accumulator.
}
