% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{contains_IM}
\alias{contains_IM}
\title{Check whether needle (value) is in haystack (vector or string)}
\usage{
contains_IM(haystack, needle)
}
\arguments{
\item{haystack}{Vector or string to search through}

\item{needle}{Value to search for}
}
\value{
Boolean
}
\description{
Equivalent of .Contains() in Insight Maker
}
\examples{
contains_IM(c("a", "b", "c"), "d") # FALSE
contains_IM(c("abcdef"), "bc") # TRUE
}
\seealso{
Other custom: 
\code{\link{\%REM\%}()},
\code{\link{expit}()},
\code{\link{filter_IM}()},
\code{\link{indexof}()},
\code{\link{length_IM}()},
\code{\link{logistic}()},
\code{\link{logit}()},
\code{\link{mod}()},
\code{\link{rbool}()},
\code{\link{rdist}()},
\code{\link{rem}()},
\code{\link{round_IM}()},
\code{\link{shuffle}()},
\code{\link{substr_i}()}
}
\concept{custom}
