% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay.R
\name{delayN}
\alias{delayN}
\title{Delay of order N}
\usage{
delayN(variable, length, order, initial = NULL)
}
\arguments{
\item{variable}{Variable to delay.}

\item{length}{Length of the delay.}

\item{order}{Order of the delay.}

\item{initial}{Value to initialize the accumulator with. If NULL, the first value of the variable to delay is used. Defaults to NULL.}
}
\value{
Delayed variable
}
\description{
Delay of order N
}
