% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay.R
\name{delayN}
\alias{delayN}
\title{Delay of order N}
\usage{
delayN(variable, length, order, initial = NULL)
}
\arguments{
\item{variable}{Variable to delay.}

\item{length}{Length of the delay.}

\item{order}{Order of the delay.}

\item{initial}{Value to initialize the accumulator with. If NULL, the first value of the variable to delay is used. Defaults to NULL.}
}
\value{
Delayed variable
}
\description{
Delay of order N, which is a more complex delay that can be used to model processes with multiple stages or orders of delay. Note that this function is only usable for simulations run with Julia.
}
\examples{
\dontshow{if (not_on_cran()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
\seealso{
Other delay: 
\code{\link{delay}()},
\code{\link{past}()},
\code{\link{smoothN}()}
}
\concept{delay}
