% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{discard}
\alias{discard}
\title{Remove variable(s) or model unit(s)}
\usage{
discard(sfm, name)
}
\arguments{
\item{sfm}{Stock-and-flow model, object of class \code{\link{sdbuildR}}.}

\item{name}{Name(s) to remove. Accepts bare symbols (e.g., \code{x}), strings, or vectors via \code{c()}. Can be variable names or model unit names.}
}
\value{
A stock-and-flow model object of class \code{\link[=sdbuildR]{sdbuildR}}
}
\description{
Remove variable(s) or model unit(s) from a stock-and-flow model. For variables, all references in flow connections and graphical function sources are also removed. A warning will be thrown if any lingering references to the removed name remain in the model.
}
\examples{
sfm <- sdbuildR() |>
 build(x, stock)
as.data.frame(sfm)

sfm <- discard(sfm, x)
as.data.frame(sfm)
}
\seealso{
\code{\link[=build]{build()}}, \code{\link[=custom_unit]{custom_unit()}}, \code{\link[=change_name]{change_name()}}
}
