% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{drop_u}
\alias{drop_u}
\title{Drop unit in equation}
\usage{
drop_u(x)
}
\arguments{
\item{x}{Variable with unit}
}
\value{
Unitless variable
}
\description{
In rare cases, it may be desirable to drop the units of a variable within an equation. Use \code{drop_unit()} to render a variable unitless. Note that units are only supported in Julia, not in R.
}
\seealso{
\code{\link[=model_units]{model_units()}}, \code{\link[=unit_prefixes]{unit_prefixes()}}, \code{\link[=u]{u()}}, \code{\link[=convert_u]{convert_u()}}
}
