% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{expit}
\alias{expit}
\title{Expit function}
\usage{
expit(x)
}
\arguments{
\item{x}{Numerical value}
}
\value{
Numerical value
}
\description{
Inverse of the logit function
}
\examples{
expit(1)
}
\seealso{
Other custom: 
\code{\link{\%REM\%}()},
\code{\link{contains_IM}()},
\code{\link{filter_IM}()},
\code{\link{indexof}()},
\code{\link{length_IM}()},
\code{\link{logistic}()},
\code{\link{logit}()},
\code{\link{mod}()},
\code{\link{rbool}()},
\code{\link{rdist}()},
\code{\link{rem}()},
\code{\link{round_IM}()},
\code{\link{shuffle}()},
\code{\link{substr_i}()}
}
\concept{custom}
