% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{export_plot}
\alias{export_plot}
\title{Save plot to a file}
\usage{
export_plot(
  pl,
  file,
  width = 3,
  height = 4,
  units = "cm",
  dpi = 300,
  font_family = ""
)
}
\arguments{
\item{pl}{Plot object.}

\item{file}{File path to save plot to, including a file extension. For plotting a stock-and-flow model, the file extension can be one of png, pdf, svg, ps, eps, webp. For plotting a simulation, the file extension can be one of png, pdf, jpg, jpeg, webp. For plotting a qgraph graph, the file extension can be one of png, pdf, svg, ps, eps, jpg, jpeg, tiff, bmp. If no file extension is specified, it will default to png.}

\item{width}{Width of image in units.}

\item{height}{Height of image in units.}

\item{units}{Units in which width and height are specified. Either "cm", "in", or "px".}

\item{dpi}{Resolution of image. Only used if units is not "px".}

\item{font_family}{Font family used for qgraph exports. For PDF/PS/EPS exports,
this is applied when the graphics device is opened.}
}
\value{
Returns \code{NULL} invisibly, called for side effects.
}
\description{
Save a plot of a stock-and-flow diagram or a simulation to a specified file path. Note that saving plots requires additional packages to be installed (see below).
}
\examples{

# Only if dependencies are installed
if (requireNamespace("DiagrammeRsvg", quietly = TRUE) &&
  requireNamespace("rsvg", quietly = TRUE)) {
  sfm <- sdbuildR("SIR")
  file <- tempfile(fileext = ".png")
  export_plot(plot(sfm), file)

  # Remove plot
  file.remove(file)
}

\dontshow{if (has_internet()) withAutoprint(\{ # examplesIf}
\dontrun{
# requires internet
# Only if suggested dependencies are installed
if (requireNamespace("htmlwidgets", quietly = TRUE) &&
  requireNamespace("webshot2", quietly = TRUE)) {
  # Requires Chrome to save plotly plot:
  sim <- simulate(sfm)
  export_plot(plot(sim), file)

  # Remove plot
  file.remove(file)
}
}
\dontshow{\}) # examplesIf}
}
\concept{simulate}
