% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{export_plot}
\alias{export_plot}
\title{Save plot to a file}
\usage{
export_plot(pl, filepath, width = 3, height = 4, units = "cm", dpi = 300)
}
\arguments{
\item{pl}{Plot object.}

\item{filepath}{File path to save plot to, including a file extension. For plotting a stock-and-flow model, the file extension can be one of png, pdf, svg, ps, eps, webp. For plotting a simulation, the file extension can be one of png, pdf, jpg, jpeg, webp. If no file extension is specified, it will default to png.}

\item{width}{Width of image in units.}

\item{height}{Height of image in units.}

\item{units}{Units in which width and heigth are specified. Either "cm", "in", or "px".}

\item{dpi}{Resolution of image. Only used if units is not "px".}
}
\description{
Save a plot of a stock-and-flow diagram or a simulation to a specified filepath.
}
\examples{
sfm <- xmile("SIR")
filepath <- tempfile(fileext = ".png")
export_plot(plot(sfm), filepath)

# Remove plot
file.remove(filepath)

\dontshow{if (not_on_cran() & interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
sim <- simulate(sfm)
export_plot(plot(sim), filepath)

# Remove plot
file.remove(filepath)
\dontshow{\}) # examplesIf}
}
\seealso{
Other simulate: 
\code{\link{ensemble}()},
\code{\link{plot.sdbuildR_ensemble}()},
\code{\link{plot.sdbuildR_sim}()},
\code{\link{sim_specs}()},
\code{\link{simulate}()},
\code{\link{solvers}()},
\code{\link{use_julia}()},
\code{\link{use_threads}()}
}
\concept{simulate}
