% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{find_dependencies}
\alias{find_dependencies}
\title{Find dependencies}
\usage{
find_dependencies(sfm, reverse = FALSE)
}
\arguments{
\item{sfm}{Stock-and-flow model, object of class sdbuildR_xmile.}

\item{reverse}{If FALSE, list for each variable X which variables Y it depends on for its equation definition. If TRUE, don't show dependencies but dependents. This reverses the dependencies, such that for each variable X, it lists what other variables Y depend on X.}
}
\value{
List, with for each model variable what other variables it depends on, or if reverse = TRUE, which variables depend on it
}
\description{
Find which other variables each variable is dependent on.
}
\examples{
sfm <- xmile("SIR")
find_dependencies(sfm)

}
\seealso{
Other build: 
\code{\link{as.data.frame.sdbuildR_sim}()},
\code{\link{as.data.frame.sdbuildR_xmile}()},
\code{\link{build}()},
\code{\link{debugger}()},
\code{\link{get_build_code}()},
\code{\link{header}()},
\code{\link{macro}()},
\code{\link{plot.sdbuildR_xmile}()},
\code{\link{print.summary.sdbuildR_xmile}()},
\code{\link{summary.sdbuildR_xmile}()},
\code{\link{xmile}()}
}
\concept{build}
