% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{get_build_code}
\alias{get_build_code}
\title{Generate code to build stock-and-flow model}
\usage{
get_build_code(sfm)
}
\arguments{
\item{sfm}{Stock-and-flow model, object of class sdbuildR_xmile.}
}
\value{
String with code to build stock-and-flow model from scratch.
}
\description{
Create R code to rebuild an existing stock-and-flow model. This may help to understand how a model is built, or to modify an existing one.
}
\examples{
sfm <- xmile("SIR")
get_build_code(sfm)

}
\seealso{
Other build: 
\code{\link{as.data.frame.sdbuildR_sim}()},
\code{\link{as.data.frame.sdbuildR_xmile}()},
\code{\link{build}()},
\code{\link{debugger}()},
\code{\link{find_dependencies}()},
\code{\link{header}()},
\code{\link{macro}()},
\code{\link{plot.sdbuildR_xmile}()},
\code{\link{print.summary.sdbuildR_xmile}()},
\code{\link{summary.sdbuildR_xmile}()},
\code{\link{xmile}()}
}
\concept{build}
