% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{get_regex_units}
\alias{get_regex_units}
\title{Get regular expressions for units in Julia}
\usage{
get_regex_units(sfm = NULL)
}
\arguments{
\item{sfm}{Stock-and-flow model, object of class sdbuildR_xmile.}
}
\value{
Named vector with regular expressions as names and units as entries
}
\description{
Get regular expressions for units in Julia
}
\examples{
get_regex_units()

}
\seealso{
Other units: 
\code{\link{convert_u}()},
\code{\link{drop_u}()},
\code{\link{get_regex_time_units}()},
\code{\link{get_units}()},
\code{\link{model_units}()},
\code{\link{u}()},
\code{\link{unit_prefixes}()}
}
\concept{units}
