% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{get_units}
\alias{get_units}
\title{View all standard units}
\usage{
get_units()
}
\value{
A character matrix with 5 columns: \code{description} (unit description),
\code{name} (unit symbol or abbreviation), \code{full_name} (full unit name),
\code{definition} (mathematical definition in terms of base units), and \code{prefix}
(logical indicating whether SI prefixes like kilo- or milli- can be applied).
Includes SI base units, derived units, CGS units, US customary units, and
custom units added by sdbuildR.
}
\description{
Obtain a data frame with all standard units in Julia's Unitful package and added custom units by sdbuildR.
}
\examples{
x <- get_units()
head(x)

}
\concept{units}
