% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{\%REM\%}
\alias{\%REM\%}
\title{Remainder operator}
\usage{
a \%REM\% b
}
\arguments{
\item{a}{Dividend}

\item{b}{Divisor}
}
\value{
Remainder
}
\description{
This is a wrapper for the rem() function. It is used to match Insight Maker's mod, which is the remainder operator.
}
\examples{
-7 \%REM\% 3
}
\seealso{
Other custom: 
\code{\link{contains_IM}()},
\code{\link{expit}()},
\code{\link{filter_IM}()},
\code{\link{indexof}()},
\code{\link{length_IM}()},
\code{\link{logistic}()},
\code{\link{logit}()},
\code{\link{mod}()},
\code{\link{rbool}()},
\code{\link{rdist}()},
\code{\link{rem}()},
\code{\link{round_IM}()},
\code{\link{shuffle}()},
\code{\link{substr_i}()}
}
\concept{custom}
