% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{indexof}
\alias{indexof}
\title{Find index of needle (value) in haystack (vector)}
\usage{
indexof(haystack, needle)
}
\arguments{
\item{haystack}{Vector or string to search through}

\item{needle}{Value to search for}
}
\value{
Index
}
\description{
Equivalent of .IndexOf() in Insight Maker
}
\examples{
indexof(c("a", "b", "c"), "b") # 2
indexof("haystack", "hay") # 1
}
