% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{indexof}
\alias{indexof}
\title{Find index of needle (value) in haystack (vector)}
\usage{
indexof(haystack, needle)
}
\arguments{
\item{haystack}{Vector or string to search through}

\item{needle}{Value to search for}
}
\value{
Index
}
\description{
Equivalent of .IndexOf() in Insight Maker
}
\examples{
indexof(c("a", "b", "c"), "b") # 2
indexof("haystack", "hay") # 1
indexof("haystack", "m") # 0
}
\seealso{
Other custom: 
\code{\link{contains_IM}()},
\code{\link{expit}()},
\code{\link{length_IM}()},
\code{\link{logistic}()},
\code{\link{logit}()},
\code{\link{rbool}()},
\code{\link{rdist}()},
\code{\link{rem}()},
\code{\link{round_IM}()}
}
\concept{custom}
