% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{logistic}
\alias{logistic}
\title{Logistic function}
\usage{
logistic(x, slope = 1, midpoint = 0, upper = 1)
}
\arguments{
\item{x}{Value}

\item{slope}{Slope of logistic function. Defaults to 1.}

\item{midpoint}{Midpoint of logistic function where the output is .5. Defaults to 0.}

\item{upper}{Maximal value returned by logistic function. Defaults to 1.}
}
\value{
f(x), where f is the logistic function
}
\description{
Logistic function
}
\examples{
logistic(0)
logistic(1, slope = 5, midpoint = 0.5)
}
\concept{custom}
