% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{logit}
\alias{logit}
\title{Logit function}
\usage{
logit(p)
}
\arguments{
\item{p}{Probability, numerical value between 0 and 1}
}
\value{
Logit
}
\description{
Logit function
}
\examples{
logit(.1)
}
\seealso{
Other custom: 
\code{\link{contains_IM}()},
\code{\link{expit}()},
\code{\link{indexof}()},
\code{\link{length_IM}()},
\code{\link{logistic}()},
\code{\link{rbool}()},
\code{\link{rdist}()},
\code{\link{rem}()},
\code{\link{round_IM}()}
}
\concept{custom}
