% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_xmile.R
\name{macro}
\alias{macro}
\title{Create, modify or remove a global variable or function}
\usage{
macro(sfm, name, eqn = "0.0", doc = "", change_name = NULL, erase = FALSE)
}
\arguments{
\item{sfm}{Stock-and-flow model, object of class sdbuildR_xmile.}

\item{name}{Name of the macro. The equation will be assigned to this name.}

\item{eqn}{Equation of the macro. A character vector. Defaults to "0.0".}

\item{doc}{Documentation of the macro. Defaults to "".}

\item{change_name}{New name for macro (optional). Defaults to NULL to indicate no change.}

\item{erase}{If TRUE, remove macro from the model. Defaults to FALSE.}
}
\value{
Updated stock-and-flow model
}
\description{
Macros are global variables or functions that can be used throughout your stock-and-flow model. \code{macro()} adds, changes, or erases a macro.
}
\examples{
# If the sigmoid() function did not exist, you could create it yourself:
sfm = xmile() \%>\%
macro("sig", eqn = "function(x, slope = 1, midpoint = .5) 1 / (1 + exp(-slope*(x-midpoint)))")

}
