% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{nonnegative}
\alias{nonnegative}
\title{Safely check whether x is less than zero}
\usage{
nonnegative(x)
}
\arguments{
\item{x}{Value}
}
\value{
x if x is greater than 0, 0 otherwise
}
\description{
If using Julia, units are preserved
}
\examples{
nonnegative(NA)
nonnegative(-1)

}
\seealso{
Other internal: 
\code{\link{clean_name}()},
\code{\link{compare_sim}()},
\code{\link{has_internet}()},
\code{\link{not_on_cran}()},
\code{\link{saveat_func}()}
}
\concept{internal}
