% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{nonnegative}
\alias{nonnegative}
\title{Safely check whether x is less than zero, preserving units}
\usage{
nonnegative(x)
}
\arguments{
\item{x}{Value}
}
\value{
x if x is greater than 0, 0 otherwise
}
\description{
Safely check whether x is less than zero, preserving units
}
\examples{
nonnegative(NA)
nonnegative(-1)

}
