% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_xmile.R
\name{plot.sdbuildR_xmile}
\alias{plot.sdbuildR_xmile}
\title{Plot stock-and-flow diagram}
\usage{
\method{plot}{sdbuildR_xmile}(x, format_label = TRUE, wrap_width = 25, center_stocks = FALSE, ...)
}
\arguments{
\item{x}{Stock-and-flow model of class sdbuildR_xmile}

\item{format_label}{If TRUE, apply default formatting to labels if labels are the same as variable names.}

\item{wrap_width}{Width of text wrapping for labels. Must be an integer.}

\item{center_stocks}{If TRUE, stocks are centered in the diagram. Defaults to FALSE.}

\item{...}{Optional arguments}
}
\value{
Stock-and-flow diagram plotted with DiagrammeR()
}
\description{
Visualise a stock-and-flow diagram using DiagrammeR. Stocks are represented as boxes. Flows are represented as arrows between stocks and/or double circles, where the latter represent what it outside of the model boundary.
}
\examples{
sfm = xmile("SIR")
plot(sfm)
}
\seealso{
\code{\link[=insightmaker_to_sfm]{insightmaker_to_sfm()}}
}
