% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plot.sdbuildR_xmile}
\alias{plot.sdbuildR_xmile}
\title{Plot stock-and-flow diagram}
\usage{
\method{plot}{sdbuildR_xmile}(
  x,
  format_label = TRUE,
  wrap_width = 20,
  center_stocks = FALSE,
  font_size = 18,
  font_family = "Georgia",
  stock_col = "#83d3d4",
  flow_col = "#f48153",
  minlen = 2,
  ...
)
}
\arguments{
\item{x}{Stock-and-flow model of class sdbuildR_xmile}

\item{format_label}{If TRUE, apply default formatting (removing periods and underscores) to labels if labels are the same as variable names.}

\item{wrap_width}{Width of text wrapping for labels. Must be an integer. Defaults to 20.}

\item{center_stocks}{If TRUE, stocks are vertically aligned in the middle of the diagram. Defaults to FALSE.}

\item{font_size}{Font size. Defaults to 18.}

\item{font_family}{Font name. Defaults to "Georgia".}

\item{stock_col}{Colour of stocks. Defaults to "#83d3d4".}

\item{flow_col}{Colour of flows. Defaults to "#f48153".}

\item{minlen}{Minimum length of edges; must be an integer. Defaults to 2.}

\item{...}{Optional arguments}
}
\value{
Stock-and-flow diagram plotted with DiagrammeR()
}
\description{
Visualize a stock-and-flow diagram using DiagrammeR. Stocks are represented as boxes. Flows are represented as arrows between stocks and/or double circles, where the latter represent what it outside of the model boundary. Hover over the stocks and flows to see their equations.
}
\examples{
sfm <- xmile("SIR")
plot(sfm)

}
\seealso{
\code{\link[=insightmaker_to_sfm]{insightmaker_to_sfm()}}, \code{\link[=xmile]{xmile()}}, \code{\link[=plot.sdbuildR_sim]{plot.sdbuildR_sim()}}

Other build: 
\code{\link{as.data.frame.sdbuildR_sim}()},
\code{\link{as.data.frame.sdbuildR_xmile}()},
\code{\link{build}()},
\code{\link{debugger}()},
\code{\link{find_dependencies}()},
\code{\link{get_build_code}()},
\code{\link{header}()},
\code{\link{macro}()},
\code{\link{print.summary.sdbuildR_xmile}()},
\code{\link{summary.sdbuildR_xmile}()},
\code{\link{xmile}()}
}
\concept{build}
