% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.simulate_stockflow}
\alias{plot.simulate_stockflow}
\title{Plot timeseries of simulation}
\usage{
\method{plot}{simulate_stockflow}(
  x,
  show_constants = FALSE,
  vars = NULL,
  palette = "Dark 2",
  colors = NULL,
  line_width = 2,
  font_family = "Times New Roman",
  font_size = 16,
  wrap_width = 25,
  showlegend = TRUE,
  animation = c("none", "time"),
  webgl = getOption("sdbuildR.webgl", default = TRUE),
  ...
)
}
\arguments{
\item{x}{Output of \code{\link[=simulate.stockflow]{simulate()}}.}

\item{show_constants}{If \code{TRUE}, include constants in plot. Defaults to \code{FALSE}.}

\item{vars}{Variables to plot. Defaults to \code{NULL} to plot all variables.}

\item{palette}{Colour palette. Must be one of hcl.pals().}

\item{colors}{Vector of colours. If \code{NULL}, the color palette will be used. If specified, will override palette. The number of colours must be equal to the number of variables in the simulation data frame. Defaults to \code{NULL}.}

\item{line_width}{Line width of the plotted trajectories. Either a single value applied to all variables, or a vector with one value per variable (like \code{colors}). Defaults to \code{2}.}

\item{font_family}{Font family. Defaults to "Times New Roman".}

\item{font_size}{Font size. Defaults to 16.}

\item{wrap_width}{Width of text wrapping for labels. Must be an integer. Defaults to 25.}

\item{showlegend}{Whether to show legend. Must be \code{TRUE} or \code{FALSE}. Defaults to \code{TRUE}.}

\item{animation}{Animation mode. Use \code{"none"} for a static plot or \code{"time"}
to cumulatively reveal trajectories over time. Defaults to \code{"none"}.}

\item{webgl}{If \code{TRUE}, render trajectories with WebGL (plotly \code{scattergl}) for
performance with many lines; if \code{FALSE}, use SVG (\code{scatter}). Defaults to
\code{getOption("sdbuildR.webgl", default = TRUE)}. Set
\code{options(sdbuildR.webgl = FALSE)} (e.g. in vignettes or dashboards, or when
a plot renders blank) to disable WebGL globally.}

\item{...}{Optional parameters}
}
\value{
Plotly object
}
\description{
Visualize simulation results of a stock-and-flow model. Plot the evolution of stocks over time, with the option of also showing other model variables.
}
\examples{
sfm <- stockflow("sir")
sim <- simulate(sfm)
plot(sim)

# The default plot title and axis labels can be changed like so:
plot(sim, main = "Simulated trajectory", xlab = "Time", ylab = "Value")

# Add constants to the plot
plot(sim, show_constants = TRUE)

# Cumulatively reveal the trajectories over time
plot(sim, animation = "time")

}
\seealso{
\code{\link[=simulate.stockflow]{simulate()}}, \code{\link[=as.data.frame.simulate_stockflow]{as.data.frame.simulate_stockflow()}}, \code{\link[=plot.simulate_stockflow]{plot.simulate_stockflow()}}
}
\concept{simulate}
