% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_xmile.R
\name{print.summary.sdbuildR_xmile}
\alias{print.summary.sdbuildR_xmile}
\title{Print method for summary.sdbuildR_xmile}
\usage{
\method{print}{summary.sdbuildR_xmile}(x, ...)
}
\arguments{
\item{x}{A summary object of class "summary.sdbuildR_xmile"}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns the summary object of class summary.sdbuildR_xmile
}
\description{
Print method for summary.sdbuildR_xmile
}
\seealso{
Other build: 
\code{\link{as.data.frame.sdbuildR_sim}()},
\code{\link{as.data.frame.sdbuildR_xmile}()},
\code{\link{build}()},
\code{\link{debugger}()},
\code{\link{find_dependencies}()},
\code{\link{get_build_code}()},
\code{\link{header}()},
\code{\link{macro}()},
\code{\link{plot.sdbuildR_xmile}()},
\code{\link{summary.sdbuildR_xmile}()},
\code{\link{xmile}()}
}
\concept{build}
