% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-sdbuildR.R
\name{print.summary_sdbuildR}
\alias{print.summary_sdbuildR}
\title{Print method for summary_sdbuildR}
\usage{
\method{print}{summary_sdbuildR}(x, ...)
}
\arguments{
\item{x}{Object of class \code{summary_sdbuildR}}

\item{...}{Ignored}
}
\value{
\code{x} invisibly
}
\description{
Print method for summary_sdbuildR
}
\concept{build}
