% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{pulse}
\alias{pulse}
\title{Create pulse function}
\usage{
pulse(start, height = 1, width = 1, repeat_interval = NULL)
}
\arguments{
\item{start}{Start time of pulse}

\item{height}{Height of pulse, defaults to 1}

\item{width}{Width of pulse in duration (i.e. time). This cannot be equal to or less than 0; to indicate an instantaneous pulse, specify the simulation step size}

\item{repeat_interval}{Interval at which to repeat pulse, defaults to NULL to indicate no repetition}
}
\value{
Interpolation function
}
\description{
This function is coming soon, but is now unusable.
}
\details{
Equivalent of Pulse() in Insight Maker
}
