% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{ramp}
\alias{ramp}
\title{Create ramp function}
\usage{
ramp(start_t_ramp, end_t_ramp, start_h_ramp = 0, end_h_ramp = 1)
}
\arguments{
\item{start_t_ramp}{Start time of ramp}

\item{end_t_ramp}{End time of ramp}

\item{start_h_ramp}{Start height of ramp, defaults to 0}

\item{end_h_ramp}{End height of ramp, defaults to 1}
}
\value{
Interpolation function
}
\description{
Equivalent of Ramp() in Insight Maker
}
