% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{ramp}
\alias{ramp}
\title{Create ramp function}
\usage{
ramp(start, finish, height = 1)
}
\arguments{
\item{start}{Start time of ramp}

\item{finish}{End time of ramp}

\item{height}{End height of ramp, defaults to 1}
}
\value{
Interpolation function
}
\description{
Equivalent of Ramp() in Insight Maker
}
