% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{rdist}
\alias{rdist}
\title{Generate random number from custom distribution}
\usage{
rdist(a, b)
}
\arguments{
\item{a}{Vector to draw sample from}

\item{b}{Vector of probabilities}
}
\value{
One sample from custom distribution
}
\description{
Equivalent of RandDist() in Insight Maker
}
\examples{
rdist(c(1, 2, 3), c(.5, .25, .25))
}
\seealso{
Other custom: 
\code{\link{\%REM\%}()},
\code{\link{contains_IM}()},
\code{\link{expit}()},
\code{\link{filter_IM}()},
\code{\link{indexof}()},
\code{\link{length_IM}()},
\code{\link{logistic}()},
\code{\link{logit}()},
\code{\link{mod}()},
\code{\link{rbool}()},
\code{\link{rem}()},
\code{\link{round_IM}()},
\code{\link{shuffle}()},
\code{\link{substr_i}()}
}
\concept{custom}
