% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{round_IM}
\alias{round_IM}
\title{Round Half-Up (as in Insight Maker)}
\usage{
round_IM(x, digits = 0)
}
\arguments{
\item{x}{Value}

\item{digits}{Number of digits; optional, defaults to 0}
}
\value{
Rounded value
}
\description{
R rounds .5 to 0, whereas Insight Maker rounds .5 to 1. This function is the equivalent of Insight Maker's Round() function.
}
\examples{
round_IM(.5) # 1
round(.5) # 0
round_IM(-0.5) # 0
round(-0.5) # 0
round_IM(1.5) # 2
round(1.5) # 2
}
\seealso{
Other custom: 
\code{\link{contains_IM}()},
\code{\link{expit}()},
\code{\link{indexof}()},
\code{\link{length_IM}()},
\code{\link{logistic}()},
\code{\link{logit}()},
\code{\link{rbool}()},
\code{\link{rdist}()},
\code{\link{rem}()}
}
\concept{custom}
