% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{round_IM}
\alias{round_IM}
\title{Convert Insight Maker's Round() function to R}
\usage{
round_IM(x, digits = 0)
}
\arguments{
\item{x}{Value}

\item{digits}{Number of digits; optional, defaults to 0}
}
\value{
Rounded value
}
\description{
This function ensures that rounding in R matches rounding in Insight Maker. R base's round() rounds .5 to 0, whereas round_IM() rounds .5 to 1.
}
\details{
Source: https://stackoverflow.com/questions/12688717/round-up-from-5/12688836#12688836
}
\examples{
round_IM(.5)
}
