% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{round_IM}
\alias{round_IM}
\title{Convert Insight Maker's Round() function to R}
\usage{
round_IM(x, digits = 0)
}
\arguments{
\item{x}{Value}

\item{digits}{Number of digits; optional, defaults to 0}
}
\value{
Rounded value
}
\description{
This function ensures that rounding in R matches rounding in Insight Maker. R base's round() rounds .5 to 0, whereas round_IM() rounds .5 to 1.
}
\details{
Source: https://stackoverflow.com/questions/12688717/round-up-from-5/12688836#12688836
}
\examples{
round_IM(.5)
}
\seealso{
Other custom: 
\code{\link{\%REM\%}()},
\code{\link{contains_IM}()},
\code{\link{expit}()},
\code{\link{filter_IM}()},
\code{\link{indexof}()},
\code{\link{length_IM}()},
\code{\link{logistic}()},
\code{\link{logit}()},
\code{\link{mod}()},
\code{\link{rbool}()},
\code{\link{rdist}()},
\code{\link{rem}()},
\code{\link{shuffle}()},
\code{\link{substr_i}()}
}
\concept{custom}
