% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{saveat_func}
\alias{saveat_func}
\title{Internal function to save dataframe at specific times}
\usage{
saveat_func(df, time_col, new_times)
}
\arguments{
\item{df}{Dataframe}

\item{time_col}{Name of the time column}

\item{new_times}{Vector of new times to save the dataframe at}
}
\value{
Dataframe with new times and interpolated values
}
\description{
Internal function used to save the dataframe at specific times in case save_at is not equal to dt in the simulation specifications.
}
\seealso{
Other internal: 
\code{\link{compare_sim}()},
\code{\link{has_internet}()},
\code{\link{nonnegative}()},
\code{\link{not_on_cran}()}
}
\concept{internal}
