% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{seasonal}
\alias{seasonal}
\title{Annual sine wave with an amplitude of 1}
\usage{
seasonal(period = 1, shift = 0)
}
\arguments{
\item{period}{Duration of wave in years}

\item{shift}{Timing of wave peak in years, defaults to 0}
}
\value{
Seasonal wave
}
\description{
This function is coming soon, but is now unusable.
}
\details{
Equivalent of Seasonal() in Insight Maker
}
