% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/past.R
\name{setup_delayN}
\alias{setup_delayN}
\title{Setup a delayN accumulator}
\usage{
setup_delayN(initial, length, order, name)
}
\arguments{
\item{initial}{Initial value for the accumulator. If NULL, the first value of the variable to delay is used. Defaults to NULL.}

\item{length}{Length of the delay.}

\item{order}{Order of the delay.}

\item{name}{Name of the accumulator.}
}
\value{
A named vector of initial values for the accumulator.
}
\description{
Internal function to set up a delayN accumulator.
}
