% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_xmile.R
\name{sim_specs}
\alias{sim_specs}
\title{Modify simulation specifications}
\usage{
sim_specs(
  sfm,
  method = "euler",
  start = "0.0",
  stop = "100.0",
  dt = "0.01",
  saveat = dt,
  adaptive = FALSE,
  seed = NULL,
  time_units = "s",
  language = "Julia",
  ...
)
}
\arguments{
\item{sfm}{Stock-and-flow model, object of class sdbuildR_xmile.}

\item{method}{Integration method. Defaults to "euler".}

\item{start}{Start time of simulation. Defaults to 0.}

\item{stop}{End time of simulation. Defaults to 100.}

\item{dt}{Timestep of solver. Defaults to 0.01.}

\item{saveat}{Timestep at which to save computed values. Defaults to dt.}

\item{adaptive}{If TRUE, an adaptive (instead of a fixed) time step is used. Defaults to FALSE.}

\item{seed}{Seed number to ensure reproducibility across runs in case of random elements. Must be an integer. Defaults to NULL (no seed).}

\item{time_units}{Simulation time unit, e.g. 's' (second). Defaults to "s".}

\item{language}{Coding language in which to simulate model. Either "Julia" (recommended) or "R". Defaults to "Julia".}

\item{...}{Optional additional parameters}
}
\value{
Updated stock-and-flow model with new simulation specifications
}
\description{
Simulation specifications are the settings that determine how the model is simulated, such as the integration method, start and stop time, and timestep. Modify these specifications for an existing stock-and-flow model.
}
\examples{
sfm = xmile("predator-prey") \%>\%
sim_specs(start = 1960, stop = 2010, dt = 0.1)
sim = simulate(sfm)
plot(sim)

# Change the simulation method to "rk4"
sfm = sfm \%>\% sim_specs(method = "rk4")

# Change the time units to "years", such that one time unit is one year
sfm = sfm \%>\% sim_specs(time_units = "years")

# To save storage but not affect accuracy, use saveat
sfm = sfm \%>\% sim_specs(saveat = 0.1, dt = 0.001)
sim = simulate(sfm)
head(as.data.frame(sim))

# Specify seed for reproducibility
sfm = sfm \%>\% sim_specs(seed = 1) \%>\%
 # Add stochastic initial condition
 build(c("predator", "prey"), eqn = "runif(1, 20, 50)")
sim1 = simulate(sfm)
sim2 = simulate(sfm)
plot(sim1)
plot(sim2)

# Removing the seed yields variation
sfm = sfm \%>\% sim_specs(seed = NULL)
sim1 = simulate(sfm)
sim2 = simulate(sfm)
plot(sim1)
plot(sim2)

# Not recommended: change the simulation language to R.
# Note that units and delay functions are not supported in R.
sfm = sfm \%>\% sim_specs(language = "R")
}
