% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate stock-and-flow model}
\usage{
simulate(
  sfm,
  keep_nonnegative_flow = TRUE,
  keep_nonnegative_stock = FALSE,
  keep_unit = TRUE,
  only_stocks = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{sfm}{Stock-and-flow model, object of class sdbuildR_xmile.}

\item{keep_nonnegative_flow}{If TRUE, keeps original non-negativity setting of flows. Defaults to TRUE.}

\item{keep_nonnegative_stock}{If TRUE, keeps original non-negativity setting of stocks Defaults to FALSE.}

\item{keep_unit}{If TRUE, keeps units of variables. Defaults to TRUE.}

\item{only_stocks}{If TRUE, only save stocks. If FALSE, auxiliaries and flows are saved using a callback function. Only applies if language is set to "Julia" in sim_specs() and no delay functions are used. Defaults to FALSE.}

\item{verbose}{If TRUE, update on progress. Defaults to FALSE.}

\item{...}{Optional arguments}
}
\value{
Object of class sdbuildR_sim, which is a list containing:
\describe{
\item{df}{Dataframe, timeseries of computed variables in the ODE}
\item{init}{Initial value of stocks}
\item{constants}{Constant parameters}
\item{script}{Simulation script}
\item{duration}{Duration of simulation}
\item{success}{If TRUE, simulation was successful. If FALSE, simulation failed.}
\item{...}{Other parameters passed to simulate}
}
}
\description{
Simulate a stock-and-flow model with simulation specifications defined by \code{sim_specs()}. If not already run, the Julia environment will first be set up with \code{use_julia()}. If any problems are detected by \code{debugger()}, the model cannot be simulated.
}
\examples{
sfm <- xmile("SIR")
sim <- simulate(sfm)
plot(sim)

# Obtain all model variables
sim <- simulate(sfm, only_stocks = FALSE)
plot(sim, add_constants = TRUE)

\dontrun{
# requires Julia setup
if (julia_status()$status == "ready") {
  # Use Julia for models with units or delay functions
  sfm <- sim_specs(xmile("coffee_cup"), language = "Julia")
  sim <- simulate(sfm)
  plot(sim)

  # Close Julia session
  use_julia(stop = TRUE)
}
}
}
\seealso{
\code{\link[=build]{build()}}, \code{\link[=xmile]{xmile()}}, \code{\link[=debugger]{debugger()}}, \code{\link[=sim_specs]{sim_specs()}}, \code{\link[=use_julia]{use_julia()}}

Other simulate: 
\code{\link{ensemble}()},
\code{\link{export_plot}()},
\code{\link{plot.sdbuildR_ensemble}()},
\code{\link{plot.sdbuildR_sim}()},
\code{\link{sim_specs}()},
\code{\link{solvers}()},
\code{\link{use_threads}()}
}
\concept{simulate}
