% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_func.R
\name{step}
\alias{step}
\title{Create step function}
\usage{
step(start, height = 1)
}
\arguments{
\item{start}{Start time of step}

\item{height}{Height of step, defaults to 1}
}
\value{
Interpolation function
}
\description{
Create a step function that jumps from zero to a specified height at a specified time, and remains at that height until the end of the simulation time.
}
\details{
Equivalent of Step() in Insight Maker
}
