% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockflow.R
\name{stockflow}
\alias{stockflow}
\title{Create a new stock-and-flow model}
\usage{
stockflow(template = NULL)
}
\arguments{
\item{template}{Name of the template to load. If \code{NULL}, an empty stock-and-flow
model will be created with default simulation parameters and a default meta.
If specified, \code{template} should be one of the available templates (case-insensitive):
\itemize{
\item \strong{logistic_model}: Population growth with carrying capacity
\item \strong{sir}: Epidemic model (Susceptible-Infected-Recovered)
\item \strong{predator_prey}: Lotka-Volterra dynamics
\item \strong{cusp}: Cusp catastrophe model
\item \strong{crielaard2022}: Eating behavior (doi: 10.1037/met0000484)
\item \strong{coffee_cup}: Temperature equilibration (Meadows)
\item \strong{bank_account}: Compound interest (Meadows)
\item \strong{lorenz}: Lorenz attractor (chaotic)
\item \strong{rossler}: Rossler attractor (chaotic)
\item \strong{vanderpol}: Van der Pol oscillator
\item \strong{duffing}: Forced Duffing oscillator
\item \strong{chua}: Chua's circuit (chaotic)
\item \strong{jdr}: Job Demands-Resources Theory as formalized in Evers et al. (under review)
}}
}
\value{
A stock-and-flow model object of class \code{\link{stockflow}}. Its structure is based
on \href{https://docs.oasis-open.org/xmile/xmile/v1.0/os/xmile-v1.0-os.html}{XML Interchange Language for System Dynamics (XMILE)}. It is a nested list, containing:
\describe{
\item{meta}{Meta-information about model. A list containing arguments listed in \code{\link[=meta]{meta()}}.}
\item{sim_settings}{Simulation specifications. A list containing arguments listed in \code{\link[=sim_settings]{sim_settings()}}.}
\item{model}{Model variables, grouped under the variable types stock, flow, aux (auxiliaries), constant, gf (graphical functions), and func (custom functions). Each variable contains arguments as listed in \code{\link[=update]{update()}}.}
}

Use \code{\link[=summary]{summary()}} to run model diagnostics, \code{\link[=as.data.frame]{as.data.frame()}} to convert to a data.frame, \code{\link[=plot]{plot()}} to visualize.
}
\description{
Initialize a stock-and-flow model of class \code{\link{stockflow}}. You can
either create an empty stock-and-flow model or load a template from the model
library.
}
\details{
Do not edit the object manually; this will likely lead to errors downstream.
Rather, use \code{\link[=meta]{meta()}}, \code{\link[=sim_settings]{sim_settings()}}, \code{\link[=update]{update()}}, and \code{\link[=custom_func]{custom_func()}} for safe manipulation.
}
\examples{
sfm <- stockflow()
summary(sfm)

\dontshow{
sfm <- sim_settings(sfm, save_at = 1)
}

# Load a template
sfm <- stockflow("lorenz")
sim <- simulate(sfm)
plot(sim)
}
\seealso{
\code{\link[=update]{update()}}, \code{\link[=meta]{meta()}}, \code{\link[=custom_func]{custom_func()}}, \code{\link[=sim_settings]{sim_settings()}}
}
\concept{build}
