% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_xmile.R
\name{summary.sdbuildR_xmile}
\alias{summary.sdbuildR_xmile}
\title{Print overview of stock-and-flow model}
\usage{
\method{summary}{sdbuildR_xmile}(object, ...)
}
\arguments{
\item{object}{A stock-and-flow model object of class \code{\link[=xmile]{sdbuildR_xmile}}}

\item{...}{Optional arguments}
}
\value{
Summary object of class \link{summary.sdbuildR_xmile}
}
\description{
Print summary of stock-and-flow model, including number of stocks, flows, constants, auxiliaries, graphical functions, macros, and custom model units, as well as simulation specifications and use of delay functions.
}
\examples{
sfm <- xmile("SIR")
summary(sfm)

}
\seealso{
\code{\link[=build]{build()}}
}
\concept{build}
