% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{unit_prefixes}
\alias{unit_prefixes}
\title{Show unit prefixes}
\usage{
unit_prefixes()
}
\value{
Matrix with unit prefixes, symbols, and power-of-ten scale
}
\description{
Show unit prefixes
}
\examples{
unit_prefixes()
}
