% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insightmaker_conv.R
\name{url_to_IM}
\alias{url_to_IM}
\title{Extract Insight Maker model XML string from URL}
\usage{
url_to_IM(URL, file)
}
\arguments{
\item{URL}{String with URL to an Insight Maker model}

\item{file}{String with file path to an Insight Maker model with suffix .InsightMaker}
}
\value{
String with Insight Maker model
}
\description{
For internal use; use \code{insightmaker_to_sfm()} to import an Insight Maker model.
}
\seealso{
\code{\link[=insightmaker_to_sfm]{insightmaker_to_sfm()}}

Other insightmaker: 
\code{\link{insightmaker_to_sfm}()}
}
\concept{insightmaker}
