% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insightmaker_conv.R
\name{url_to_IM}
\alias{url_to_IM}
\title{Extract Insight Maker model XML string from URL}
\usage{
url_to_IM(URL, filepath_IM, directory)
}
\arguments{
\item{URL}{String with URL to an Insight Maker model}

\item{filepath_IM}{String with file path to an Insight Maker model with suffix .InsightMaker}

\item{directory}{Directory to save output in, optional}
}
\value{
String with Insight Maker model
}
\description{
Extract Insight Maker model XML string from URL
}
