% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_setup.R
\name{use_julia}
\alias{use_julia}
\title{Set up Julia environment}
\usage{
use_julia(stop = FALSE, JULIA_HOME = NULL, force = FALSE)
}
\arguments{
\item{stop}{If TRUE, stop active Julia session. Defaults to FALSE.}

\item{JULIA_HOME}{Path to Julia installation. Defaults to NULL to locate Julia automatically.}

\item{force}{If TRUE, force Julia setup to execute again.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Create Julia environment to simulate stock-and-flow models. \code{use_julia()} looks for a Julia installation, and if found, will set up a Julia project enviroment. In most cases, this will require installing some packages. The Julia project environment is located in the inst directory of the sdbuildR package ("Project.toml").
}
\details{
Keep in mind that the installation can take around 5-20 minutes. In every R session, \code{use_julia()} needs to be run once (which is done automatically in \code{simulate()}), which can take around 30 seconds.
}
\examples{
\dontrun{ # requires Julia setup
if (JuliaConnectoR::juliaSetupOk()) {
  # Start Julia session or set up environment (first time use)
  use_julia()

  # Stop Julia session
  use_julia(stop = TRUE)
}
}
}
\seealso{
Other simulate: 
\code{\link{ensemble}()},
\code{\link{export_plot}()},
\code{\link{plot.sdbuildR_ensemble}()},
\code{\link{plot.sdbuildR_sim}()},
\code{\link{sim_specs}()},
\code{\link{simulate}()},
\code{\link{solvers}()},
\code{\link{use_threads}()}
}
\concept{simulate}
