% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_setup.R
\name{use_julia}
\alias{use_julia}
\title{Set up Julia environment}
\usage{
use_julia(stop = FALSE, force = FALSE)
}
\arguments{
\item{stop}{If TRUE, stop active Julia session. Defaults to FALSE.}

\item{force}{If TRUE, force Julia setup to execute again.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Start Julia session and activate Julia environment to simulate stock-and-flow models. \code{use_julia()} looks for a Julia installation and an instantiated Julia environment specifically for sdbuildR. This can be set up with \code{install_julia_env()}. For more guidance, please see \href{https://kcevers.github.io/sdbuildR/articles/julia-setup.html}{this vignette}.
}
\details{
In every R session, \code{use_julia()} needs to be run once (which is done automatically in \code{simulate()}), which can take around 30-60 seconds.
}
\examples{
\dontrun{
# requires Julia setup
if (julia_status()$status == "ready") {
  # Start a Julia session and activate the Julia environment for sdbuildR
  use_julia()

  # Stop Julia session
  use_julia(stop = TRUE)
}
}
}
\seealso{
Other julia: 
\code{\link{install_julia_env}()},
\code{\link{julia_status}()}
}
\concept{julia}
