% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_setup.R
\name{use_julia}
\alias{use_julia}
\title{Set up Julia environment}
\usage{
use_julia(stop = FALSE, JULIA_HOME = NULL, force = FALSE)
}
\arguments{
\item{stop}{If TRUE, stop active Julia session. Defaults to FALSE.}

\item{JULIA_HOME}{Path to Julia installation. Defaults to NULL to locate Julia automatically.}

\item{force}{If TRUE, force Julia setup to execute again.}
}
\description{
Create Julia environment to simulate stock-and-flow models. \code{use_julia()} looks for a Julia installation, and will install Julia as well as some required packages if not found. A Julia environment is created in the inst directory of the sdbuildR package.
}
\details{
Keep in mind that the installation can take around 5-15 minutes. In every R session, \code{use_julia()} needs to be run once (which is done automatically in simulate()), which can take around 30 seconds.
}
\examples{
if (FALSE) {
  # Start Julia session
  # or install Julia and set up environment (first time use)
  use_julia()

  # Stop Julia session
  use_julia(stop = TRUE)
}
}
\seealso{
Other simulate: 
\code{\link{ensemble}()},
\code{\link{export_plot}()},
\code{\link{plot.sdbuildR_ensemble}()},
\code{\link{plot.sdbuildR_sim}()},
\code{\link{sim_specs}()},
\code{\link{simulate}()},
\code{\link{solvers}()},
\code{\link{use_threads}()}
}
\concept{simulate}
