% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_xmile.R
\name{xmile}
\alias{xmile}
\title{Create a new stock-and-flow model}
\usage{
xmile(name = NULL)
}
\arguments{
\item{name}{Name of the template to load. If NULL, an empty stock-and-flow model will be created with default simulation parameters and a default header.}
}
\value{
Stock-and-flow model of class sdbuildR_xmile.
}
\description{
Initialize an empty stock-and-flow model of class sdbuildR_xmile. You can either create an empty stock-and-flow model or load a template from the model library.
}
\examples{
sfm = xmile()
summary(sfm)

# Load a template
sfm = xmile("Lorenz")
sim = simulate(sfm)
plot(sim)
}
\seealso{
\code{\link[=build]{build()}}
}
