% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_xmile.R
\name{xmile}
\alias{xmile}
\title{Create a new stock-and-flow model}
\usage{
xmile(name = NULL)
}
\arguments{
\item{name}{Name of the template to load. If NULL, an empty stock-and-flow model will be created with default simulation parameters and a default header. If specified, the name should be one of "logistic_model", "SIR", "predator-prey", "Crielaard2022", "coffee_cup", "bank_account", "Lorenz", "Rossler", "vanderPol", "Duffing", "Chua", "spruce_budworm".}
}
\value{
Stock-and-flow model of class sdbuildR_xmile.
}
\description{
Initialize an empty stock-and-flow model of class sdbuildR_xmile. You can either create an empty stock-and-flow model or load a template from the model library.
}
\examples{
sfm <- xmile()
summary(sfm)

\dontshow{
sfm <- sim_specs(sfm, save_at = 1)
}

# Load a template
sfm <- xmile("Lorenz")
sim <- simulate(sfm)
plot(sim)
}
\seealso{
\code{\link[=build]{build()}}

Other build: 
\code{\link{as.data.frame.sdbuildR_sim}()},
\code{\link{as.data.frame.sdbuildR_xmile}()},
\code{\link{build}()},
\code{\link{debugger}()},
\code{\link{find_dependencies}()},
\code{\link{get_build_code}()},
\code{\link{header}()},
\code{\link{macro}()},
\code{\link{plot.sdbuildR_xmile}()},
\code{\link{summary.sdbuildR_xmile}()}
}
\concept{build}
